Add padding in onboarding screen
